pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["dmkael <tanmaar@gmail.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
validators = "^0.28.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
